Make Garply build tree usable by clients
We can generate a GarplyConfig file in the build tree of Garply
so that clients do not need to install it before trying to build
a package against it. When set up, this means that we can do

$ mkdir build-garply && cd build-garply
$ cmake <pathtogarplysrcdir>
$ make
...
$ mkdir build-client && cd build-client
$ cmake -DCMAKE_PREFIX_PATH=/path/to/build-garply <pathtoclientsrc>

This is helpful for many development tasks and workflows, but relies
on setting include and other paths appropriately.

Add generation of Config and Targets files for Garply build tree.
Recursion to subdirectories is moved before this generation, as
creation of the Targets file for the Build tree requires the exports
set to be present.

Make build tree header path (for garply_version.h) PRIVATE so that
it does not appear in client include paths (it is only needed when
building garply, *not* when using).

Rename GarplyConfig.cmake.in template variables for header and
lib paths so these can be set in main script without interfering
with the CMAKE_INSTALL_XXX variables directly. Set the new variables
as required for Build/Install trees before calling the
configure_package_config_file function.

The added complexity of the generation is due to the very different
structures of the build/install trees. This complexity can be
reduced by full use of imported targets and usage requirements. Here,
the targets themselves automatically set and propagate paths, so the
GarplyConfig file can be identical for the build and install trees.